refactor: 🚚 rename firestore functions to firebase functions
because they are not a part of firebase
