+++ client/src/lib/firebase/firebase-functions-types.ts
@@ -1,0 +1,1 @@
+export * from "../../../../functions/src/firebase-functions-types";
