--- client/src/lib/firebase/firestore-functions-types.ts
@@ -1,1 +1,0 @@
-export * from "../../../../functions/src/firestore-functions-types";
