+++ functions/src/firebase-functions-types.ts
@@ -1,5 +1,7 @@
 export type LobbyRequest = { code: string };
 
+export type LobbyCreationResponse = { code: string };
+
 export function isLobbyRequest(data: unknown): data is LobbyRequest {
   // will only return true if the data is an object with a code property and string
   return data != null && typeof data === "object" && "code" in data && typeof (data as LobbyRequest).code === "string";
